small int ei - refactoring + backwards compatibility
